+++ test/expected/custom-loader/a.js
@@ -44,10 +44,16 @@
 /* 0 */
 /***/ function(module, exports, __webpack_require__) {
 
-	
-	console.log(1);
+	__webpack_require__(1);
+	console.log("foo");
 
-	foo
+
+
+/***/ },
+/* 1 */
+/***/ function(module, exports, __webpack_require__) {
+
+	// removed by extract-text-webpack-plugin
 
 /***/ }
 /******/ ]);