+++ test/fixtures/custom-loader/a.js
@@ -1,3 +1,3 @@
-
+require('./a.css');
 console.log(1);