+++ test/fixtures/custom-loader/foo.js
@@ -2,5 +2,7 @@
 
 module.exports = function(source) {
   this.cacheable && this.cacheable();
-  return source + 'foo';
+  return source
+    .replace('bar', 'boo')
+    .replace(1, '\"foo\"');
 };